TLC for the Shift model
* Moved the managers to their module, and merged into one
* Renamed to volunteers. That seems to be our word to use.
