+++ scheduler/models.py
@@ -2,50 +2,17 @@
 
 from datetime import timedelta
 
-from dateutil.parser import parse
-
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils import timezone
 from django.utils.formats import localize
+
 from django.utils.translation import ugettext_lazy as _
 
-from places.models import Country, Area, Place, Region
+from .managers import EnrolmentManager, ShiftManager, NeedManager
 
-from .managers import EnrolmentManager
 
 ## KEPT FROM OLD FOR NEW MODEL
-
-
-class NeedManager(models.Manager):
-    def at_location(self, location):
-        return self.get_queryset().filter(location=location)
-
-    def at_place(self, place):
-        return self.get_queryset().filter(location__place=place)
-
-    def in_area(self, area):
-        return self.get_queryset().filter(location__place__area=area)
-
-    def in_region(self, region):
-        return self.get_queryset().filter(location__place__area__region=region)
-
-    def in_country(self, country):
-        return self.get_queryset().filter(
-            location__place__area__region__country=country)
-
-    def by_geography(self, geo_affiliation):
-        if isinstance(geo_affiliation, Location):
-            return self.at_location(geo_affiliation)
-        elif isinstance(geo_affiliation, Place):
-            return self.at_place(geo_affiliation)
-        elif isinstance(geo_affiliation, Area):
-            return self.in_area(geo_affiliation)
-        elif isinstance(geo_affiliation, Region):
-            return self.in_region(geo_affiliation)
-        elif isinstance(geo_affiliation, Country):
-            return self.in_country(geo_affiliation)
-
 
 class Location(models.Model):
     """
@@ -275,7 +242,7 @@ class RecurringEvent(models.Model):
     name = models.CharField(max_length=255)
     description = models.TextField(max_length=20000, blank=True)
     weekday = models.IntegerField(choices=WEEKDAYS, null=False)
-    slot_amount = models.IntegerField(verbose_name=_(u'number of needed volunteers'))
+    needed_volunteers = models.IntegerField(verbose_name=_(u'number of needed volunteers'))
     start_time = models.CharField(verbose_name=_('Starting time'), max_length=5, validators=[time_validator])
     end_time = models.CharField(verbose_name=_('Ending time'), max_length=5)
     first_date = models.DateTimeField(verbose_name=_('First occurence'))
@@ -298,66 +265,22 @@ class RecurringEvent(models.Model):
         self.save(*args, **kwargs)
 
 
-# This "second" view allows to create/publish the shifts for the X coming days out of the recurring events
-# On REST side, it should be possible to request all events from the template for a specified day,
-# return them to the front-end which will publish them and/or modify them via PUTs statements
-
-class ShiftManager(NeedManager):
-
-    def create_datetime_from_day_and_time_string(self, day, time_str):
-        return parse(time_str, default=day)
-
-    def create_shift_from_event_and_day(self, event, day):
-        shift = Shift(task=event.task,
-                      workplace=event.workplace,
-                      name=event.name,
-                      description=event.description,
-                      slot_amount=event.slot_amount,
-                      start_time=self.create_datetime_from_day_and_time_string(day, event.start_time),
-                      end_time=self.create_datetime_from_day_and_time_string(day, event.end_time))
-        # shift.save() # see if it should be done here later
-        return shift
-
-    def create_shifts_for_facility_and_day(self, facility, day):
-        events = RecurringEvent.objects.get_events_for_facility_and_day(facility, day)
-        shifts = []
-        for event in events:
-            shifts.append(self.create_shift_from_event_and_day(event, day))
-        return shifts
-
-
-class OpenShiftManager(ShiftManager):
-    def get_queryset(self):
-        now = timezone.now()
-        qs = super(OpenShiftManager, self).get_queryset()
-        return qs.filter(end_time__gte=now)
-
-
 class Shift(models.Model):
-    facility = models.ForeignKey("TBD.Facility", verbose_name=_(u''), help_text=_(u''))
+    """
+    A shift. Happens at a time and place, and hopefully has many volunteers attached to it.
+    """
     task = models.ForeignKey("scheduler.Task", verbose_name=_(u''), help_text=_(u''))
-    workplace = models.ForeignKey("scheduler.Workplace", verbose_name=_(u''), help_text=_(u''))
-    helpers = models.ManyToManyField('accounts.UserAccount', through='Enrolment', related_name='shifts')
+    workplace = models.ForeignKey("places.Workplace", verbose_name=_(u''), help_text=_(u''))
+    volunteers = models.ManyToManyField(
+        'accounts.UserAccount', through='scheduler.Enrolment', related_name='shifts')
     # to set only if created from template. But really useful? Maybe just more complexity
     # origin_event = models.ForeignKey("RecurringEvent", null=True, verbose_name=_(u''), help_text=_(u''))
-    name = models.CharField(max_length=255)
-    description = models.TextField(max_length=20000, blank=True)
-    slot_amount = models.IntegerField(verbose_name=_(u'number of needed volunteers'))
+    needed_volunteers = models.IntegerField(verbose_name=_(u'number of needed volunteers'))
     start_time = models.DateTimeField(verbose_name=_('starting time'), db_index=True)
     end_time = models.DateTimeField(verbose_name=_('ending time'), db_index=True)
     created_at = models.DateTimeField(verbose_name=_('created at'), auto_now_add=True)
     updated_at = models.DateTimeField(verbose_name=_('updated at'), auto_now=True)
     published_at = models.DateTimeField(verbose_name=_('published at'), null=True)
     cancelled_at = models.DateTimeField(verbose_name=_('cancelled at'), null=True)
-    # shift_slug = models.CharField(max_length=255)
 
-    shifts = NeedManager()
-    open_shifts = OpenShiftManager()
-
-    # @property
-    # def slugify(self):
-    #     return '{}-{}:{}'.format(self.start_time.isoformat(), self.start_time, self.end_time.hours, self.end_time.minutes)
-
-    def save(self, *args, **kwargs):
-        # self.shift_slug = self.slugify
-        self.save(*args, **kwargs)
+    objects = ShiftManager()
